+++ lib/hexapdf/cli/optimize.rb
@@ -66,6 +66,7 @@ module HexaPDF
       end
 
       def execute(in_file, out_file) #:nodoc:
+        check_output_file(out_file)
         HexaPDF::Document.open(in_file, decryption_opts: {password: @password}) do |doc|
           optimize_page_tree(doc)
           apply_optimization_options(doc)